+++ src/containers/BootStrap.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import { UserContext } from 'context/UserContext'
+import AuthProvider from 'context/Authentication'
 import { useQuery } from '@apollo/client'
 import { ME_QUERY } from 'modules/users/queries'
 import {
@@ -23,8 +23,6 @@ interface LocationState {
 const Public = () => <h3>This is public content</h3>
 const Protected = () => <h3>This is hidden content</h3>
 
-
-
 const Bootstrap: React.FC = ({}) => {
   const { loading, error, data } = useQuery(ME_QUERY)
 
@@ -38,11 +36,11 @@ const Bootstrap: React.FC = ({}) => {
     return <PublicRoutes />
   }
   return (
-    <UserContext.Provider value={me}>
+    <AuthProvider user={me}>
       <Router>
         <MainLayout />
       </Router>
-    </UserContext.Provider>
+    </AuthProvider>
   )
 }
